+++ CHANGELOG.md
@@ -7,5 +7,9 @@
 
 ## Other
 -->
+## Changes
+- Remove download location redirection from `Settings > Downloads`
+- Moved cache related stuff from `Settings > Advanced` to `Settings > Data and storage`
+
 ## Other
 - Migrate `RawQueries#librayQuery` to use SQLDelight (`GetLibraryManga`)